fix(classPage): address 6 code review findings from PR 271
- Reuse compareStudentNames in classPageAdapter.ts (DRY)
- Fix surfaceState race condition: return loading when adapterResult is null
  even if dataset reports ready (HIGH — real bug)
- Extract 4 compute* helpers + ClassPageError type to useClassPageData.helpers.ts
  (keeps useClassPageData.ts under 500-line threshold at 404 lines)
- Add test asserting loading status when pipeline hasn't produced output
- Return null from mockClassPage component (fixes React undefined-render warning)
- Remove 3 empty duplicate JSDoc blocks in classPageAdapter.zod.spec.ts
- Correct JSDoc in StudentAveragesTableCard.tsx to match actual useMemo deps

Lint: 0 errors, 0 warnings. Tests: 1436 passing (1 new). TSC: clean.
